[Refactor] client -> infrastructure/external로 이동
